requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic-settings==2.0.3
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
# PostgreSQL driver for Supabase Postgres support
psycopg2-binary==2.9.9
# Supabase Python client
supabase==2.3.4
# Email validation for Pydantic models
email-validator==2.1.0
# FastAPI-MCP for AI assistant integration
fastapi-mcp==0.3.0
# OpenAI client for Groq API compatibility
openai==1.54.3
